refactoring recieve steps
